feat: implement language toggle with i18n context and cookie persistence (#27)
