+++ src/blocks/LanguageToggleBlock/LanguageToggleBlock.tsx
@@ -4,10 +4,12 @@ type Language = "en" | "pt";
 
 type LanguageToggleBlockProps = {
   currentLanguage: Language;
+  onLanguageChange: (next: Language) => void;
 };
 
 export function LanguageToggleBlock({
   currentLanguage,
+  onLanguageChange,
 }: LanguageToggleBlockProps) {
   return (
     <div className={styles.container}>
@@ -16,13 +18,13 @@ export function LanguageToggleBlock({
         role="group"
         aria-label="Language selector"
       >
-
         <button
           type="button"
           className={`${styles.button} ${
             currentLanguage === "pt" ? styles.active : ""
           }`}
           aria-pressed={currentLanguage === "pt"}
+          onClick={() => onLanguageChange("pt")}
         >
           PT
         </button>
@@ -33,6 +35,7 @@ export function LanguageToggleBlock({
             currentLanguage === "en" ? styles.active : ""
           }`}
           aria-pressed={currentLanguage === "en"}
+          onClick={() => onLanguageChange("en")}
         >
           EN
         </button>
